move(ScheduleTimeResponse): ScheduleTimeResponse를 service위치로 이동
