+++ backend/src/main/java/com/woowacourse/momo/service/meeting/MeetingService.java
@@ -7,8 +7,8 @@ import com.woowacourse.momo.domain.meeting.Meeting;
 import com.woowacourse.momo.domain.meeting.MeetingRepository;
 import com.woowacourse.momo.domain.schedule.Schedule;
 import com.woowacourse.momo.domain.schedule.ScheduleRepository;
-import com.woowacourse.momo.domain.schedule.dto.ScheduleTimeResponse;
 import com.woowacourse.momo.service.meeting.dto.MeetingResponse;
+import com.woowacourse.momo.service.schedule.dto.ScheduleTimeResponse;
 import java.time.LocalDate;
 import java.util.Comparator;
 import java.util.List;
